GCW-3650 Create separate file for en translations
