+++ src/i18n/en.json
@@ -1,0 +1,8 @@
+{
+  "Home": "Home",
+  "Chats": "Chats",
+  "Offers": "Offers",
+  "Login": "Login",
+  "Log out": "Log out",
+  "Get 4-digit SMS code": "Get 4-digit SMS code"
+}
