+++ src/index.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import ReactDOM from "react-dom";
 import App from "App";
 import * as serviceWorker from "serviceWorker";
-import "i18n";
+import "i18n/i18n";
 
 ReactDOM.render(<App />, document.getElementById("root"));
